pyproject.toml
[project]
name = "arc-solver"
version = "0.1.0"
description = "System 1/System 2 ARC-AGI-2 Solver"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "aiofiles>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

